[Bug] fixes creature division of step speed (#676)
• Fixed NPCs speed (were walking too fast)
• Fixed the speed of the imbuements (they were bent)
• Fixed the speed of the Gods
• Fixed creature step speed being divided by 2
• Some others things related to speed

Fixes #431

+++ src/io/iologindata.cpp
@@ -287,7 +287,7 @@ bool IOLoginData::loadPlayer(Player* player, DBResult_ptr result)
   player->loginPosition.z = result->getNumber<uint16_t>("posz");
 
   player->addPreyCards(result->getNumber<uint64_t>("prey_wildcard"));
-  player->addTaskHuntingPoints(result->getNumber<uint16_t>("task_points"));
+  player->addTaskHuntingPoints(result->getNumber<uint64_t>("task_points"));
 
   player->lastLoginSaved = result->getNumber<time_t>("lastlogin");
   player->lastLogout = result->getNumber<time_t>("lastlogout");
